refactor: 그룹 좌표 타입 정리 및 ddl 정합, 엔티티 매핑에 MapStruct 도입
- 위도/경도를 BigDecimal에서 Double로 변경 (ddl.sql이 DECIMAL(10,7)에서
  DOUBLE로 바뀐 것과 일치)
- 그룹 테이블/컬럼명을 최신 ddl.sql과 통일
  ( → dining_group, group_member.group_id → dining_group_id)
- GroupEntity ↔ NewGroup/Group 간 수동 매핑 코드를 GroupMapper(MapStruct)로 대체
- GroupIntegrationTest의 불필요한 @ActiveProfiles(test) 제거

+++ api/src/main/java/com/gm/api/controller/group/dto/request/GroupCreateRequest.java
@@ -1,6 +1,5 @@
 package com.gm.api.controller.group.dto.request;
 
-import java.math.BigDecimal;
 import java.time.LocalTime;
 import java.util.UUID;
 
@@ -27,12 +26,12 @@ public record GroupCreateRequest(
         @NotNull
         @DecimalMin("-90")
         @DecimalMax("90")
-        BigDecimal latitude,
+        Double latitude,
 
         @NotNull
         @DecimalMin("-180")
         @DecimalMax("180")
-        BigDecimal longitude,
+        Double longitude,
 
         @NotNull
         @Positive
