+++ core/src/main/java/com/gm/core/domain/group/model/NewGroup.java
@@ -1,6 +1,5 @@
 package com.gm.core.domain.group.model;
 
-import java.math.BigDecimal;
 import java.time.LocalTime;
 import java.util.UUID;
 
@@ -20,8 +19,8 @@ public record NewGroup(
         UUID ownerUserId,
         String name,
         String locationAddress,
-        BigDecimal latitude,
-        BigDecimal longitude,
+        Double latitude,
+        Double longitude,
         int searchRadiusM,
         LocalTime recommendationTime,
         int maxMemberCount
